Suppress FxCop about Uri
This one I don't think should apply universally to the test library, so
I'm just adding the suppression locally. The reason I think this might
be better to do locally is that the motivation for the rule is that the
Uri class provides better encapsulation. I agree. Had I been writing
code that'd be passing variables around, I think that the rule should
still be in effect.

This test, on the other hand, currently uses a string literal, so I
think this actually makes the test more readable.