Create reservation in repository
There's SO much to say about this commit. It's bigger than I usually
like them to be, and it took a good deal more than 5 minutes to produce
it. This isn't unusual in the beginning of a project, but can still feel
frustrating if one's not prepared for it.

Im using the outside-in TDD style that I like so much. Contrary to the
way many people often interpret BDD, outside-in TDD starts at the
boundary of the application, but it doesn't have to stay there. In this
commit, I dropped a level of abstraction and now instead test against a
class instead of against the HTTP API. People would classify that as a
unit test rather than a boundary test, so would tend to place it in a
different file, perhaps even in a different test library. I'm more
interested in driving the behaviour of the various features, so you
could say that I organise my code along the lines of features instead of
according to technical details.

ReservationsController does the least amount of work I could get away
with. Notice that it just saves hard-coded values!

I had to add the [ApiController] attribute to ReservationsController to
make the boundary test pass.

While working to make all tests pass, I managed to make the new test
pass, only to see the other test of /reservations fail. It required some
troubleshooting to identify the problem. It might be worth writing about
as well.

In the Startup class, I decided to do the simplest thing that could
possibly work, and that was a private Null Object implementation. I also
decided to try to use the built-in DI Container in ASP.NET Core. After
much heated twitter debate, I understand that I'm biased. The way I
usually do it works and has advantages, but perhaps there are advantages
on the other side of the fence. I'm trying it on for size, at least.

I chose to name the database abstraction with the Repository suffix. I'm
not too fond of doing that these days, but I still pick that name
because I think that most people will be able to easily pick up on the
purpose of it.

The Reservation domain model is more than 50% tool-generated.