Save reservation in SQL Server
This commit isn't test-driven, since we can regard
SqlReservationsRepository as a Humble Object. The only branching in the
Create method was driven by static code analysis, and added with a
refactoring.

The change to using a SQL Server-based repository triggered some changes
to Startup. Apart from the obvious addition of a configurable connection
string, notice also that this required ConfigureServices to be an
instance method. It also prompted the removal of the private
NullRepository class.

I also added a connection string to the appsettings.json file, but I
left the value blank. The value might contain sensitive information, so
shouldn't be checked into source control.

Finally, I had to fix the boundary test, because I don't want it to run
against a database. Not yet, anyway.