Use empty string for null name
I debated back and forth with myself how to deal with a null string, but
I decided to translate null to "". According to Postel's law, we should
be liberal in what we accept. The name isn't really required for
decision-making.

Another point to ponder is the hidden branching in the test. I'm usually
not fond of branching in tests, but the alternative is to write a whole
new test that's near-identical.