Accept reservation if allocation is possible
This is beginning to look more appropriate for a configuration with pure
communal tables.

The code is beginning to look complex. The cyclomatic complexity is 5,
so there's still a bit of head room there, but the line count is 23,
which is uncomfortably close to the 24 line maximum.

I'm also not convinced that null inputs should cause exceptions. If I
were to follow Postel's law, I should return false on null reservations.
That would require a deliberate design decision, though, so I think that
there ought to be tests for those scenarios.