Filter reservations after the day in question
This implementation is still not correct, because the filter only allows
through other reservations on the exact same date and time (to the
tick). More test cases are required to improve the situation.

The cyclomatic complexity of the WillAccept method is now at 7 (just
acceptable), but the line count is 27, which is too much. The number of
variables is 7, which is also at the limit.