Inject MaitreD into ReservationsController
This is the first step to making the restaurant more configurable. The
plan is to put both seating duration and the tables into a configuration
file.

There are two options for configuring ReservationsController:
1. Inject a MaitreD object
2. Inject configuration values (seating duration and tables)

I went with the first option because one could imagine that in the
future, a MaitreD might need even more sophisticated configuration. For
example, one might want to initialise it with opening hours, or a
calendar with holidays where the restaurant is closed. If one went with
the second option, one would have to change the ReservationsController
constructor every time the MaîtreD constructor changed.

Notice that a MaitreD is Concrete Dependency. It's more like a Parameter
Object than a 'traditional' dependency.