Read existing reservation
This is quite a commit, but is basically triggered by the addition of a
single parametrised test.

There's a few shortcuts that should be addressed as soon as possible. As
Kent Beck suggests in TDD by Example, I keep notes on the side with
issues that outght to be addressed. These include:

- There's a suppression of C# null reference flow analysis (a !
  operator) in ReservationsController.Get
- new Guid(id) in ReservationsController.Get may fail to parse