Return 404 Not Found for non-guid id
The decision to return 404 Not Found when requesting an ID which isn't a
GUID may seem like an odd design decision, but I find it
incontrovertible that a resource like /reservations/foo doesn't exist.

If we consider the design as a RESTful design instead of RPC, then
clients are supposed to follow links (like the previous
ReadSuccessfulReservation test does). If clients do that, the URLs
themselves should be considered opaque identifiers without semantic
content.