Return error message on insufficient capacity
Notice that I consider it okay to add more assertions to an existing
test. When I'm not quite so keen on is that the test code spans too many
lines. I ought to shorten it, but I'll wait until I get to that part in
the narrative where that fits.

The actual error message could be more elaborate, but to keep the sample
code at a printable line width, I'll consider that message sufficient.