Enable deletion of a reservation
I dicussed with a former colleague who's been writing security-sensitive
code for ten years whether I should add any security features to an
operation like this.

Over a nice lunch we agreed that as always, it's a trade-off. On the one
hand, you want to prevent hostile agents from being able to delete other
peoples' reservations. On the other hand, you don't want to scare users
away by making them create an account and log in.

If we assume that the no-log-in requirement weighs more heavily than a
hostile delete, there's not much more you can do. To delete the
reservation, you need the ID, and since it's a GUID, it's virtually
impossible to guess. Assuming that the address is shared only over
SSL/TLS, even a man-in-the-middle attack shouldn't reveal the ID.

It seems, however, like a good time to add an email feature to the
service. When a user makes changes, the service should send an email to
the address in the reservation.