Add PUT capability to reservations resource
Quite a lot of functionality is missing from this commit. If one had to
integrate this particular commit with master before addressing the
issues, hiding it behind a feature flag would be appropriate.

No input validation (or rather: parsing) takes place, so the
implementation is vulnerable to malformed content. Additionally, the
update goes straight to the database, bypassing all business rules.