Parse reservation ID on PUT
As usual, the argument is that a URL like /reservations/baz represents a
missing resource, and thus 404 Not Found is the appropriate response.

I chose to write the test as a unit test (as opposed to a HTTP-API-based
test). It's a trade-off.

The advantage of writing HTTP-API-based tests is that they're less
vulnerable to implementation details in how requests and responses are
represent in ASP.NET Core. For instance, if we later change the response
from NotFoundResult to NotFoundObjectResult, this new test would fail,
because it's not asserting at the HTTP API.

On the other hand, trying to drive this particular test scenario via the
HTTP API would require the test to 'hack' the URL construction, which
I'm trying to minimise to keep the REST API evolvable.