Use correct ID on PUT
If given conflicting IDs, the ID in the URL takes precedence. The
assumption is that for a proper REST API, the client is supposed to
follow URLs. Thus, the address that a client PUTs to should be valid
(or, at least, have been valid at the time it was given).

The current solution (to null out dto.Id) is an example of obscure code.
It should be refactored.