Don't allow PUT on absent resource
This seems to go against Postel's law, because it's not as liberal as it
could be. It would be technically possible to turn the operation into a
database CREATE if we discover that the reservation isn't present.

I can think of two scenarios where this might happen. The first is a
race condition where a previous DELETE request deleted the resource.
This might not even have to be much of a tight race, since a client
could legally have bookmarked the URL for the reservation in question.
Thus, a PUT could conceptually arrive hours (or longer) after the
DELETE. I think it'd be confusing that a PUT would enable a reservation
to spring back to life after it was deleted, so the API disallows that.

Another scenario is if a client doesn't follow links, but instead
constructs URLs based on a template. While Postel's law indicates that a
method should be liberal in what it accepts, Hyrum's law implies that
the API should disallow such misuse.