Ask MaitreD if PUT is legal
It may be a bit of a stretch to use a single test case as a cause to
bring in the entire MaitreD machinery, but while techniques such as
Devil's Advocate exist, it's worth keeping in mind that ultimately,
testing isn't an adversarial process.

There's still a bug in the implementation code, though. If the PUT
attempts to change a reservation without moving it to a new day, it's
going to count doubly. If close to sold out, this could lead to a valid
update being rejected.