Refactor SpyPostOffice to record observations
This is in preparation to being able to distinguish between various
operations, such as create, update, and delete.

This follows from the principle of not editing both tests and production
code in the same commit. Not that I manage to strictly follow that rule,
but it's good to keep in mind. Here, I started writing a test that would
check that an email is sent when a reservation is deleted, but then I
realised that with the current implementation of SpyPostOffice, I
wouldn't be able to distinguish between reservations that were added
because of a create operation and one that was added because of a delete
operation. So I stashed my work and made this commit instead.

One might notice that this design of SpyPostOffice starts to look like
reinvention of reusable Test Double libraries (AKA 'Mock libraries'). If
I planned on pursuing this path for good, I might consider pulling in a
reusable library like Moq. I don't, however.