Send email when reservation is updated
Even according to the Rule of Three, refactoring of SmtpPostOffice is
now warranted. I plan on adding a fourth method to the IPostOffice
interface, however, so I think that it'd be safer to refactor after
that.

The Put method on ReservationsController is getting unacceptable long.
Something should be done to it soon.