Refactor SmtpPostOffice
One could improve it even more, but since this is already an
improvement, commit it.

It turns out that while the code passes all tests, it fails an ad-hoc
exploratory test because of a defect in SqlReservationsRepository: the
ReadReservation method has swapped email and name. In Lean software
development, bugs are stop-the-line issues, so this now takes top
priority, although it's going to mean a major detour.