Introduce Email and Name wrappers
These classes wrap email and name strings. This makes it clearer to
programmers if they're accidentally trying to use an email where they
should supply a name, or vice versa. The cause of this particular change
is a defect in SqlReservationsRepository. The bug is still there, but is
now quite conspicuous. While I already knew about the defect in
ReadReservation, I hadn't realised until I made this change that the
same defect is also present in ReadReservations.

Most of the implementation code of Email and Name are generated by
Visual Studio, as is also the case with the changes to
Reservation.Equals; I deleted that method and asked Visual Studio to
generate it again. The `==` implementations are done by hand, though, as
is also the case with the ToString overrides.

It's tempting to introduce implicit conversions between Name and string,
and between Email and string. It's technically possible, but would defy
the purpose. The entire point of the wrappers is that they should NOT be
easily interchangeable.