Introduce Build solution
The plan is that Build.sln will also contain integration tests. Since
these tests are going to take more time to run, they shouldn't be
included in the normal solution that developers work with on a daily
basis.

Another alternative would be to write integration tests against a
compiled binary of Restaurant.sln. This would, however, make it more
painful to make simple refactorings (such as renames) that would also
affect the integration tests.