Change master database name to lower case
This is a typical build-server-troubleshooting commit. On Azure DevOps,
the build fails with the message that connection to the database timed
out.

On my machine, the master database is, indeed, called `master`, with a
lower-case `m`, but the tests still succeeded with an upper-case `M`. My
hypothesis is that this is because the default database collation on my
local machine is case insensitive (it is), but that the Azure DevOps
collation may not be.

If that hypothesis is correct, then the build should succeed when I push
this commit.