Add conversion to Reservation DTO
Why not an instance method on Reservation? That would couple Reservation
to ReservationDto. There's already a coupling from ReservationDto to
Reservation (which is okay), but a cyclic dependency is a cause of much
grief.

I'd prefer an extension method (ToDto), but that'd require a new static
class just for that single purpose. As an expirement, I instead tried to
use an explicit cast.

I'm not sure that that's a good idea. First of all, it seems 'clever',
which isn't positively meant. It's likely to confuse newcomers to the
code base. Also, it seems that Visual Studio interprets a cast as
something that can potentially fail; at least, it infers that the result
of the conversion is ReservationDto? (i.e. a nullable reference),
instead of ReservationDto.

I'm going to try to refactor to an extension method next, but if you
never see a commit to that effect, that implies that that experiment
didn't improve things.