Enlist Post in lightweight transaction
Notice that this makes the Build quite slow. I've somewhat arbitrarily
decided to run the race-condition test for 30 seconds. The problem with
testing concurrency is that it isn't deterministic. I must say, though,
that before I added the transaction scope, the test failed immediately,
so I'm fairly confident that it reproduces a real problem.

Using a TransactionScope around a set of injected dependencies is
actually a leaky abstraction. For less leaky abstractions, see the
discussion in the book.

This commit also disables parallel tests for the integration tests. This
is necessary because each test sets up and tears down a database of a
constant name. It might be possible to bring back parallelisation by
using a database name per test collection, but I've actually never tried
that.