Add feature flag for the calendar feature
It annoys me that it should be so complicated to do this. I blame the
built-in DI Container for most of the trouble. The HomeController
shouldn't have to take a `CalendarFlag` as a dependency - a simple
`bool` would have been sufficient. That's a concession to the DI
Container, which doesn't support value types.

To make matters even worse, I couldn't even make the wrapper a struct,
again because the DI Container doesn't allow value types.