Return Uri from UrlBuilder.BuildAbsolute
This seems redundant, since all callers immediately call ToString on the
returned Uri object.

However, my next move is goig to be the introduction of an extension
method to create a link, and I don't think it's a good idea to have such
an extension method hanging off of any arbitrary string. It makes sense
to define it for Uri objects, though.