Link to next day
This commit contains the same type of bug as in previous commits. The
code also comes with much duplication, and the AddLinks method is now so
long that it no longer fits on a single screen on my computer (I always
code on my laptop with no external screen).

I'm going to flush out the bugs first, and then refactor. Both the test
code and the implementation code needs a loving hand.