Navigate to specific years
This test actually doesn't flush out any bugs, but I expect similar
tests for months and days to do so. I decided to first add these test
cases for years to get a feeling for how to write the navigation code.
It's not pretty at the moment, but I'll improve it once it works the way
it's supposed to work.