Navigate to specific dates
These new test cases demonstrated the bugs in the day calendar's
navigation links. I think that the basic navigation functionality is now
done, BUT the tests take a long time to run. Particularly those that
navigate to dates decades removed in time are problematic.

The problem is that day navigation happens one day at a time, and it
always starts on the current date. Between July 10, 2020 and February 9,
2050 are 10.805 days, which implies 10,806 HTTP requests. That takes
some time, and at the moment, the database isn't even involved!

A fairly straightforward optimisation would be to first navigate at the
year level, then the month level, and then jump back to the date. This
requires additional links on the calendars, though, which I intend to
add.