Refactor Table to use Visitor
This seems like the most redundant complication one could come up with,
but it's a preparation for being able to keep track of reservations
associated with tables as well.

As Kent Beck writes
(https://twitter.com/kentbeck/status/250733358307500032):

"for each desired change, make the change easy (warning: this may be
hard), then make the easy change"

This commit represents (part of) the work of making the change easy.