Schedule occurrence for each distinct date
This commit pulls out one of the big guns: FsCheck. I'd been attempting
to do standard example-driven development of the Schedule feature for
days, and couldn't figure out a good way to do this in an incremental
manner. I do think, on the other hand, that I have a good plan for how
to do this with property-based testing.

FsCheck isn't the most C#-friendly library around (it's written in F#),
but I hope that the test implementation (ScheduleImp) at least looks
enough like a regular parametrised test that the average C# developer
understands what's going on there.

The assertion is currently weaker than the implementation, but I'm going
to dispense with the Devil's Advocate technique here. Just because a
technique like Devil's Advocate exists doesn't mean that you have to
always use it. I think that it'd make things harder than they have to
be, so I'm skipping it.

I plan to add more assertions to the method. Each assertion will, in
essence, be another property.

I actually like the current query-syntax-based implementation, but I may
not stick to it as the method evolves.