Use configured tables when scheduling
The implementation of Schedule still cheats because it 'knows' that no
tests (except for the degenerate test where there's no reservations)
have surplus tables in the configuration. It takes advantage of that
knowledge to zip the two collections, which is really not appropriate.

This commit introduces the Capacity property for the Table class. It's
the same as the Seats property I earlier deleted. It turns out that that
information is useful after all (not that surprising, if I may say so).
I did decide to name it Capacity instead of Seats, though, so that it's
clearer that a Table's Capacity isn't necessarily the same as its
remaining seats (which is another property that I wouldn't be surprised
to see soon).