Add Accept method to Table
This is another relaxation of the API intended to facilitate the next
assertion of MaitreD.Schedule. While I expect to be able to write that
assertion by extracting a collection of reservations from both standard
and communal tables, I don't want to expose a IEnumerable<Reservation>
as a read-only property on Table, because that makes the distinction
between standard and communal tables more implicit.