Include overlapping reservations in each occurrenc
I had to pull out some heavy artillery for this last stretch. It took me
longer than I would have liked, including a couple of hours of yak
shaving.

Some of the FsCheck data generation is way too ad-hoc to my liking, but
as the saying goes: perfect is the enemy of the good.

While I like the elegance of the query expression in the Schedule
method, I'm a bit concerned about its performance characteristics. I
haven't done a formal analysis of it, but intuitively it strikes me that
it may have quadratic complexity. Time will tell...

Another issue with this commit, and the process that lead to it, is that
there's some repetition between the production code and the test code,
when it comes to checking overlapping reservations. Again, perfect is
the enemy of the good, but I'd definitely be keen to learn about a
better way.