Return MaximumPartySize
The implementation is incorrect, but it passes all tests. I plan to
write an assertion of the Time property as well, and then one or two
examples of how the day's entries change when I add reservations to the
system.

At that point I plan to simply delegate to MaitreD.Schedule, which is
already well-tested.