Made ReservationsVisitor a public class
Why not simply expose an IEnumerable<Reservation> property on Table?
That might be conventient, but I'm a bit concerned that it'll bedim that
there's two distinct types of tables. A naive client developer might
just look at an IEnumerable<Reservation> and incorrectly infer that if
there's only one reservation, then it's a standard table. For that
reason, I'm keeping the Accept method around.