Add signature query parameter to generated URLs
At the moment, the value is just "foo", which obviously isn't a proper
signature, but comes next.

For consistency's sake, some of the other methods should probably get
the same behaviour, but I here decide to be lazy and only implement the
behaviour in the single method currently used in the code base.

I may return to this later, once I've the functionality implemented how
I'd like it.