Add restaurants array to home resource
This is the first small step towards refactoring towards a multi-tenant
solution. The intent is to fill that array with a list of restaurants to
pick from.

This feature ought really to be developed behind a feature flag, but I
think that a more important point to show here is how to introduce it
without breaking compatibility. In other words, existing clients
shouldn't break. I'll use the existing unit test suite as an indication
of whether or not I succeed in doing that. Existing tests shouldn't
break.