Rename Settings to Options
This seams to be more in accordance with standard naming for ASP.NET
configuration classes. I wasn't aware of this apparent naming convention
when I created the Settings directory.

The reason I'm doing this now is that I anticipate having to make some
changes to the configuration classes, and I thought that it'd be better
to first address the naming issue, which frankly has been annoying me
since I became aware of it.