Rename namespaces
Sometimes, during development, you arrive at new insights. One of the
small insights triggered by working with multi-tenancy is that I need a
class to hold information about each restaurant, such as its name. The
most natural name to use for such a type would be `Restaurant`.

The code analysis rule CA1724, however, rightly protests against such a
type name, since it would conflict with the namespace part of the same
name.

Since this is a service, and not a published library upon which other
developers depend, changing the namespace isn't a breaking change.