Add lookup of ID based on restaurant name
This implies that the restaurant name must be unique. If so, why not
just use that as an ID?

There's a couple of reasons behind that design decision. First, one may
think that this uniqueness constraint is already strange, and it may be,
but here is the reasoning behind it. The current design is based on the
idea that when you go to the home resource, you see an enumeration of
all available restaurants. If the API should be used to display a user
interface to a human, the way you'd typically ask a human to select a
restaurant is via its name. One could think of other designs that might
be more appropriate, but that's my thinking here. I admit that it may be
a weakness of this code base not being actual production code that has
to make money, but I hope that I'm allowed to take a few shortcuts.

If restaurant names are supposed to be unique, then why not just use
those strings as IDs? Restaurant names may change over time, and we
don't want a simple rename to break clients. IDs are part of the URLs
honoured by the API, so changing the ID would break client that might
wish to edit or delete their existing reservations.