Add tenant column to database schema
While this commit doesn't look as though it's driven by tests, it is.
It's just that I'm not checking the test in, because it's not yet
passing. I do anticipate that making that test pass will require a
series of edits that I find it most expedient to commit as a series of
smaller commits.

You may wonder why the new column is an INT column with no constraints.
In the future, one could easily envision a design where restaurants are
configured using another table, and the schema would then have a foreign
key constraint from the restaurants table to the RestaurantId column.
For now, however, the restaurants are configured in appsettings.json,
and it's the programmer's responsibility that a restaurant exists with
the appropriate ID.

Notice that this commit hard-codes the restaurant ID as 1, effectively
making it impossible to reserve at any other restaurant than Hipgnosta.
While the system is beginning to look like a multi-tenant system, it
still only supports the original restaurant Hipgnosta, which gets
'grandfathered' in, so to speak.