Add Create overload taking restaurant ID to repo
Like the previous commit, this one is also driven by a test - in fact,
the same test that also drove the previous commit. And like the previous
commit, this test is still failing, so isn't checked in yet.

Notice that while this commit involves changes to unit test, all the
tests modified are the ones that interact the Fake database. None of the
HTTP-based tests are affected.

This system is a REST API, so breaking changes are only those changes
that affect clients of the REST API. The rest is essentially
implementation details. The story could have been different if the
shipping system was a software library or framework.