Accept reservation to new restaurant
This commit is larger than I'm comfortable with. If this had been a pull
request, I'd probably complain, but still review it. It's near the limit
of what I'd consider a fair size of a pull request. Still, it actually
doesn't represent more than about half a day of work, so if this commit
had been the only change to review, I would have found it acceptable.

While I had to edit some unit tests, notice that none of the HTTP-based
tests broke. I consider this to still fit the requirement that no
breaking changes must be introduced into a deployed system.

Some of the methods are growing too large. Once I'm done refactoring to
multitenancy, I plan to run Visual Studio analytics and systemetically
go through all the methods that have grown too large.