Rename SelfHostedService to SelfHostedApi
It seems to me that this class now more represents the API as such, more
than it represents a service. Perhaps this distinction seems artificial,
but perhaps it becomes clearer with the next commit.

The integration test project has its own self-hosted API, and I should
probably align the names there... On the other hand, if the names are
similar, it might mislead developers who don't notice that they're
actually two different classes with the same name.