Separate legacy behaviour in its own class
It just dawned on me that this might be a better way to model legacy
versus current behaviour. Like the LegacyApi class itself, tests that
use it should be edited with the utmost care, if at all.

It's worth noting that the entire code base already contains two classes
that encapsulate client behaviour: SelfHostedApi and its counterpart in
the integration test project. Should this duplication be addressed?

While I'll invoke the Rule of Three, it's worth considering that two
instances are just one duplication from three.

Some commercial REST APIs come with an 'official' SDK for popular
programming platforms. While not part of this code base, one could
envision that this particular REST API might come with a .NET client
SDK, a Java client SDK, a JavaScript client SDK, a Python client SDK,
and so on.

If designed right, a client SDK could (and should) be used to test the
API itself (in the spirit of Consumer-Driven Contracts). Thus, one .NET
client SDK could be used for automated testing.

If so, one could keep an old version of such a client SDK around to
check that the system introduces no breaking changes.