Rename `service` variable to `api`
Now that the classes that encapsulate the client behaviours and
self-hosting mechanism are called XyzApi it makes sense to also call the
variables `api`. This also reflects that this is how the tests interact
with the API; the objects reflect the API, so to speak.

While this commit touches tests that exercise legacy behaviour, it
doesn't constitute a breaking change in the system. It is, after all, a
renaming of a variable. More formally, though, you get an inclination
that these edits of test code doesn't represent breaking changes because
only the test code changes. The production code is untouched by this
commit.

(This isn't a guarantee that breaking changes might not sneak in later.
While it doesn't happen in this commit, you could imagine that you could
edit tests in one commit in such a way that you weaking the assertions.
This might allow breaking changes to go unnoticed in a future commit.)

This represents a useful heuristic: Don't refactor tests and production
code at the same time. Refactor production code but leave tests
untouched, or refactor tests and leave production code untouched.