Check whether maitreD is null
This is quite an edge case. Somehow, the restaurant should be gone,
while the reservation is still in the database of reservations.

I suppose this might happen if a restaurant suddenly goes out of
business, and you want to remove it from the list of restaurants, but
you still want to retain the reservations in the database because you
might want to run a custom job on the data to let the people know that
their reservation is now void.