Check authorisation for particular restaurant
Just because a user is in the "MaitreD" role doesn't mean that he or she
should be granted access to just any restaurant's schedule. Thus, the
incoming claim should contain a restaurant ID that matches the
restaurant ID in the URL.

Why not simply detect the restaurant in question based on the incoming
security context? Why put the restaurant ID in the URL in the first
place? The motivation is that this design enables a user to act on
behalf of more than one restaurant.

You could argue that this is a breaking change, since it requires a
claim that it didn't previously require. Whether or not this really is a
breaking change depends on how JWTs are issued. This is outside the
scope of the book, and hence this code base, but a typical
implementation of federated authentication and authorisation is that the
relying party (i.e. this REST API) relies on tokens issued by a service
that the organisation also controls. Tokens are typically exchanged by
such a service after having been presented with some other proof of
identity, such as a user name and password, or API key, or another JWT
for that matter. If so, that issuer could start issuing the "restaurant"
claim to existing identities that used the original restaurant. In that
case, no clients need break.

What purpose does the AccessControlClass serve? Couldn't
ScheduleController just access its User property? It can't because it
doesn't derive from ControllerBase, and I'd like to keep it that way
because it makes unit testing much harder.

Had I been using Pure DI instead of the built-in DI Container, I'd just
have injected the list of IDs directly into the Controller, but last
time I tried to explicitly configure a Controller with ASP.NET's
built-in DI Controller, I couldn't get it to work. The otherwise
redundant class works around this issue.