Redirect clients from legacy URLs
Whlie well-behaved clients may bookmark URLs, it's okay to respond to
requests with a 301 Moved Permanently response. Clients are expected to
follow redirects, so this isn't considered a breaking change.

I'm not entirely happy about having to hack together an absolute URL in
the SigningUrlHelper, but I couldn't find a way to make ASP.NET produce
an absolute URL.