Add test for requesting schedule with bogus roles
By commenting out the [Authorize(Roles = "MaitreD")] attribute on
ScheduleController that no tests actually tested whether the
functionality activated by that attribute actually works.

In many cases, I'd consider a test for the presence of an attribute to
be close to redundant, but security is too important to not cover by
tests, and there can often be subtle misunderstandings because a
developer's mental model of what an attribute does doesn't match what
the framework actually does.