Renamed legacy Get methods
I vacillated between `LegacyGet` and `GetLegacy`, but chose `LegacyGet`
because `GetLegacy` sounds like it's a method that gets a legacy for the
caller.

As this commit shows, this change flushed out some unexpected behaviour
on the part of ASP.NET. This is most likely the reason for the behaviour
that also puzzled me that I mentioned in commit
8322b07144eae7cb220eb37e479b644bb6695501. As far as I can tell,
ASP.NET's UrlHelper keeps track of a dictionary of 'ambient values'.
This seems to include all the incoming values for the current request.
Thus, if the incoming HTTP request included a URL fragment for
restaurantId, this becomes one of the ambient values - at least when
linking within the same controller. This might explain why linking
"next" and "previous" within CalendarController works without the
restaurant ID being explicitly passed around.

Apparently, when creating "next" and "previous" links from a schedule,
these ambient values are no longer present, and URL generation then
fails... sometimes. It's still not quite clear to me why this sometimes
work, while a simple rename of a completely unrelated method makes it
fail.

It was, however, a lack of good API design that the LinkToPeriod method
didn't require restaurantId as an explicit parameter.