Add IClock and implementations
Since the system time is non-deterministic, it pays to get it under
control. In this code base, I've actually managed to code around that
non-determinism, even in unit tests, but both for the benefit of future
testability, as well as for repeatability based on logs, it's beneficial
to be able to log all non-deterministic actions. See
https://blog.ploeh.dk/2020/04/06/repeatable-execution-in-c for more
information.