Move const string out of method body
This may seem a bit arbitrary, because in some of the methods they're
still included in the body, while in other methods they're private const
strings.

Truth be told, this refactoring is motivated by the 80/24 rule. Now that
I'm done with important feature work, I'm taking the time to clean up.
One way I do that is by calculating code metrics and looking for methods
longer than 24 lines of code. The good news is that there's only ten of
those in the production code base (there's certainly more in the test
code base, and I don't want to ignore those, either).

The ReadReservations method is one of these methods. Before this change
it was 33 lines of code.

You could argue (with some success, I think) that while 33 lines of
code, that's still a fine method. Unless you're writing code on your
phone, it most likely fits on a single screen. Even though I'm
programming on a small laptop, it easily fits on my screen as well. You
could continue the argument to conclude that this proves that the 80/24
rule is stupid. I'm not even going to counter that argument.

The point of the rule of small methods is to institute a LIMIT to method
size. It doesn't have to be 24 lines of code. It could be 30 or 40 if
that fits your organisation better. For this code base, I've
deliberately chosen an extreme position to show that even within narrow
constraints, you can still write humane code.